Removed the spacy language model download from the requirements
The deploy build was failing due to a memory error, and I cannot figure out how to increase the memory limit yet.
So first I'm just trying taking this huge download out to see if it will build.
Requirements.txt probably isn't the right place for a model download anyway.